implementacao da funcionalidade de pagamento

+++ car-rental/src/main/java/com/fiap/techchallenge/carrental/entity/Cliente.java
@@ -38,7 +38,7 @@ public class Cliente {
    private SexoEnum sexo;
 
    @Column(nullable = false, unique = true)
-   private int CPF;
+   private long CPF;
 
    @Column(nullable = false)
    private int telefone;
